+++ src/main/java/com/dnd/dndbattle/services/BattleArmyCreationService.java
@@ -4,5 +4,5 @@ import com.dnd.dndbattle.domain.Soldier;
 import com.dnd.dndbattle.model.BattleArmy;
 
 public interface BattleArmyCreationService extends CRUDService<BattleArmy> {
-    BattleArmy createOrUpdateBattleArmy(Integer armyId, Soldier soldier, Integer playerId, int size);
+    BattleArmy createOrUpdateBattleArmy(Long armyId, Soldier soldier, Integer playerId, int size);
 }
