+++ src/main/java/com/dnd/dndbattle/services/CRUDService.java
@@ -5,9 +5,9 @@ import java.util.List;
 public interface CRUDService<T> {
     List<?> listAll();
 
-    T getById(Integer id);
+    T getById(Long id);
 
     T saveOrUpdate(T t);
 
-    void deleteById(Integer id);
+    void deleteById(Long id);
 }
